Restructure crate
